refactor: reorganize module structure for better maintainability
Phase 1 Refactoring:
- Rename core/command → core/process with cleaner API
  - spawn_background_process(), kill(), exists() functions
  - Use as process::spawn_background_process() for clarity
- Extract cmd_* functions from main.rs → core/commands module
  - Reduced main.rs from 286 to 92 lines
  - Better separation of concerns
- Unify error handling with GhostError type
  - Replace Box<dyn Error> with type-safe errors
  - Centralized error definitions
- Rename core module → app to avoid conflicts with Rust's core
  - ghost::app::process instead of ghost::core::process
  - More idiomatic module naming

All tests pass, clippy clean, examples updated.
