+++ src/app/mod.rs
@@ -1,0 +1,4 @@
+pub mod commands;
+pub mod error;
+pub mod process;
+pub mod storage;
