--- src/core/mod.rs
@@ -1,2 +1,0 @@
-pub mod command;
-pub mod storage;
