+++ src/lib.rs
@@ -1,1 +1,1 @@
-pub mod core;
+pub mod app;
